MAINT: Convert dataset output corresponding to model input to float dtype
Necessary so LRFinder loads it up with the right datatype.

+++ echofilter/dataset.py
@@ -192,7 +192,7 @@ class TransectDataset(torch.utils.data.Dataset):
         if self.transform_post is not None:
             sample = self.transform_post(sample)
 
-        input = np.expand_dims(sample['signals'], 0)
+        input = np.expand_dims(sample['signals'], 0).astype(np.float32)
         return input, sample
 
     def __len__(self):
